Games import - null pointer if new job deleted

+++ src/main/java/totalizator/app/controllers/rest/admin/jobs/AdminJobEditRestController.java
@@ -56,6 +56,11 @@ public class AdminJobEditRestController {
 
 	@RequestMapping( method = RequestMethod.DELETE, value = "/{jobTaskTypeId}/{jobTaskId}" )
 	public void delete( final @PathVariable int jobTaskTypeId, final @PathVariable int jobTaskId ) {
+
+		if ( jobTaskId == 0 ) {
+			return;
+		}
+
 		jobTaskService.delete( jobTaskId );
 	}
